AC-845 split SQLiteStore hub and monitor slices
