+++ autocontext/src/autocontext/storage/sqlite_store.py
@@ -4,21 +4,20 @@ import json
 import sqlite3
 from collections.abc import Sequence
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, cast
+from typing import Any, cast
 
 from autocontext.storage.bootstrap_schema import bootstrap_core_schema, default_migrations_dir
 from autocontext.storage.row_types import GenerationMetricsRow, RunRow
 from autocontext.storage.sqlite_migrations import apply_python_migration_files
-
-if TYPE_CHECKING:
-    from autocontext.monitor.types import MonitorAlert, MonitorCondition
+from autocontext.storage.sqlite_store_hub import SQLiteHubStoreMixin
+from autocontext.storage.sqlite_store_monitoring import SQLiteMonitorStoreMixin
 
 SQLITE_BUSY_TIMEOUT_MS = 5_000
 AgentOutputBatch = tuple[str, str]
 AgentRoleMetricBatch = tuple[str, str, int, int, int, str, str]
 
 
-class SQLiteStore:
+class SQLiteStore(SQLiteHubStoreMixin, SQLiteMonitorStoreMixin):
     def __init__(self, db_path: Path) -> None:
         self.db_path = db_path
         self.db_path.parent.mkdir(parents=True, exist_ok=True)
@@ -1168,474 +1167,3 @@ class SQLiteStore:
             if row is None or row["best_score"] is None:
                 return None
             return float(row["best_score"])
-
-    # ---- Research Hub metadata (AC-267) ----
-
-    def _parse_hub_session_row(self, row: dict[str, Any]) -> dict[str, Any]:
-        result = dict(row)
-        result["metadata"] = self._parse_json_field(result.pop("metadata_json", "{}"), {})
-        result["shared"] = bool(result.get("shared", 0))
-        return result
-
-    def upsert_hub_session(
-        self,
-        session_id: str,
-        *,
-        owner: str | None = None,
-        status: str | None = None,
-        lease_expires_at: str | None = None,
-        last_heartbeat_at: str | None = None,
-        shared: bool | None = None,
-        external_link: str | None = None,
-        metadata: dict[str, Any] | None = None,
-    ) -> None:
-        existing = self.get_hub_session(session_id)
-        merged_owner = owner if owner is not None else (str(existing["owner"]) if existing is not None else "")
-        merged_status = status if status is not None else (str(existing["status"]) if existing is not None else "active")
-        merged_lease = lease_expires_at if lease_expires_at is not None else (
-            str(existing["lease_expires_at"]) if existing is not None else ""
-        )
-        merged_heartbeat = last_heartbeat_at if last_heartbeat_at is not None else (
-            str(existing["last_heartbeat_at"]) if existing is not None else ""
-        )
-        merged_shared = shared if shared is not None else (bool(existing["shared"]) if existing is not None else False)
-        merged_external_link = external_link if external_link is not None else (
-            str(existing["external_link"]) if existing is not None else ""
-        )
-        merged_metadata = metadata if metadata is not None else (
-            dict(existing["metadata"]) if existing is not None else {}
-        )
-
-        with self.connect() as conn:
-            conn.execute(
-                """
-                INSERT INTO hub_sessions(
-                    session_id, owner, status, lease_expires_at, last_heartbeat_at,
-                    shared, external_link, metadata_json
-                )
-                VALUES (?, ?, ?, ?, ?, ?, ?, ?)
-                ON CONFLICT(session_id) DO UPDATE SET
-                    owner = excluded.owner,
-                    status = excluded.status,
-                    lease_expires_at = excluded.lease_expires_at,
-                    last_heartbeat_at = excluded.last_heartbeat_at,
-                    shared = excluded.shared,
-                    external_link = excluded.external_link,
-                    metadata_json = excluded.metadata_json,
-                    updated_at = strftime('%Y-%m-%dT%H:%M:%fZ', 'now')
-                """,
-                (
-                    session_id,
-                    merged_owner,
-                    merged_status,
-                    merged_lease,
-                    merged_heartbeat,
-                    1 if merged_shared else 0,
-                    merged_external_link,
-                    json.dumps(merged_metadata),
-                ),
-            )
-
-    def get_hub_session(self, session_id: str) -> dict[str, Any] | None:
-        with self.connect() as conn:
-            row = conn.execute(
-                "SELECT * FROM hub_sessions WHERE session_id = ?",
-                (session_id,),
-            ).fetchone()
-            if row is None:
-                return None
-            return self._parse_hub_session_row(dict(row))
-
-    def list_hub_sessions(self) -> list[dict[str, Any]]:
-        with self.connect() as conn:
-            rows = conn.execute(
-                "SELECT * FROM hub_sessions ORDER BY updated_at DESC"
-            ).fetchall()
-            return [self._parse_hub_session_row(dict(row)) for row in rows]
-
-    def heartbeat_hub_session(self, session_id: str, *, last_heartbeat_at: str, lease_expires_at: str | None = None) -> None:
-        existing = self.get_hub_session(session_id)
-        if existing is None:
-            self.upsert_hub_session(
-                session_id,
-                last_heartbeat_at=last_heartbeat_at,
-                lease_expires_at=lease_expires_at or "",
-            )
-            return
-        self.upsert_hub_session(
-            session_id,
-            owner=str(existing["owner"]),
-            status=str(existing["status"]),
-            lease_expires_at=lease_expires_at if lease_expires_at is not None else str(existing["lease_expires_at"]),
-            last_heartbeat_at=last_heartbeat_at,
-            shared=bool(existing["shared"]),
-            external_link=str(existing["external_link"]),
-            metadata=dict(existing["metadata"]),
-        )
-
-    def _parse_hub_package_row(self, row: dict[str, Any]) -> dict[str, Any]:
-        result = dict(row)
-        result["tags"] = self._parse_json_field(result.pop("tags_json", "[]"), [])
-        result["metadata"] = self._parse_json_field(result.pop("metadata_json", "{}"), {})
-        return result
-
-    def save_hub_package_record(
-        self,
-        *,
-        package_id: str,
-        scenario_name: str,
-        scenario_family: str,
-        source_run_id: str,
-        source_generation: int,
-        title: str,
-        description: str,
-        promotion_level: str,
-        best_score: float,
-        best_elo: float,
-        payload_path: str,
-        strategy_package_path: str,
-        tags: list[str],
-        metadata: dict[str, Any] | None = None,
-        created_at: str = "",
-    ) -> None:
-        with self.connect() as conn:
-            conn.execute(
-                """
-                INSERT INTO hub_packages(
-                    package_id, scenario_name, scenario_family, source_run_id, source_generation,
-                    title, description, promotion_level, best_score, best_elo,
-                    payload_path, strategy_package_path, tags_json, metadata_json, created_at
-                )
-                VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-                ON CONFLICT(package_id) DO UPDATE SET
-                    scenario_name = excluded.scenario_name,
-                    scenario_family = excluded.scenario_family,
-                    source_run_id = excluded.source_run_id,
-                    source_generation = excluded.source_generation,
-                    title = excluded.title,
-                    description = excluded.description,
-                    promotion_level = excluded.promotion_level,
-                    best_score = excluded.best_score,
-                    best_elo = excluded.best_elo,
-                    payload_path = excluded.payload_path,
-                    strategy_package_path = excluded.strategy_package_path,
-                    tags_json = excluded.tags_json,
-                    metadata_json = excluded.metadata_json,
-                    updated_at = strftime('%Y-%m-%dT%H:%M:%fZ', 'now')
-                """,
-                (
-                    package_id,
-                    scenario_name,
-                    scenario_family,
-                    source_run_id,
-                    source_generation,
-                    title,
-                    description,
-                    promotion_level,
-                    best_score,
-                    best_elo,
-                    payload_path,
-                    strategy_package_path,
-                    json.dumps(tags),
-                    json.dumps(metadata or {}),
-                    created_at,
-                ),
-            )
-
-    def get_hub_package_record(self, package_id: str) -> dict[str, Any] | None:
-        with self.connect() as conn:
-            row = conn.execute(
-                "SELECT * FROM hub_packages WHERE package_id = ?",
-                (package_id,),
-            ).fetchone()
-            if row is None:
-                return None
-            return self._parse_hub_package_row(dict(row))
-
-    def list_hub_package_records(self) -> list[dict[str, Any]]:
-        with self.connect() as conn:
-            rows = conn.execute(
-                "SELECT * FROM hub_packages ORDER BY created_at DESC"
-            ).fetchall()
-            return [self._parse_hub_package_row(dict(row)) for row in rows]
-
-    def _parse_hub_result_row(self, row: dict[str, Any]) -> dict[str, Any]:
-        result = dict(row)
-        result["tags"] = self._parse_json_field(result.pop("tags_json", "[]"), [])
-        result["metadata"] = self._parse_json_field(result.pop("metadata_json", "{}"), {})
-        return result
-
-    def save_hub_result_record(
-        self,
-        *,
-        result_id: str,
-        scenario_name: str,
-        run_id: str,
-        package_id: str | None,
-        title: str,
-        best_score: float,
-        best_elo: float,
-        payload_path: str,
-        tags: list[str],
-        metadata: dict[str, Any] | None = None,
-        created_at: str = "",
-    ) -> None:
-        with self.connect() as conn:
-            conn.execute(
-                """
-                INSERT INTO hub_results(
-                    result_id, scenario_name, run_id, package_id, title,
-                    best_score, best_elo, payload_path, tags_json, metadata_json, created_at
-                )
-                VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-                ON CONFLICT(result_id) DO UPDATE SET
-                    scenario_name = excluded.scenario_name,
-                    run_id = excluded.run_id,
-                    package_id = excluded.package_id,
-                    title = excluded.title,
-                    best_score = excluded.best_score,
-                    best_elo = excluded.best_elo,
-                    payload_path = excluded.payload_path,
-                    tags_json = excluded.tags_json,
-                    metadata_json = excluded.metadata_json,
-                    updated_at = strftime('%Y-%m-%dT%H:%M:%fZ', 'now')
-                """,
-                (
-                    result_id,
-                    scenario_name,
-                    run_id,
-                    package_id,
-                    title,
-                    best_score,
-                    best_elo,
-                    payload_path,
-                    json.dumps(tags),
-                    json.dumps(metadata or {}),
-                    created_at,
-                ),
-            )
-
-    def get_hub_result_record(self, result_id: str) -> dict[str, Any] | None:
-        with self.connect() as conn:
-            row = conn.execute(
-                "SELECT * FROM hub_results WHERE result_id = ?",
-                (result_id,),
-            ).fetchone()
-            if row is None:
-                return None
-            return self._parse_hub_result_row(dict(row))
-
-    def list_hub_result_records(self) -> list[dict[str, Any]]:
-        with self.connect() as conn:
-            rows = conn.execute(
-                "SELECT * FROM hub_results ORDER BY created_at DESC"
-            ).fetchall()
-            return [self._parse_hub_result_row(dict(row)) for row in rows]
-
-    def _parse_hub_promotion_row(self, row: dict[str, Any]) -> dict[str, Any]:
-        result = dict(row)
-        result["metadata"] = self._parse_json_field(result.pop("metadata_json", "{}"), {})
-        return result
-
-    def save_hub_promotion_record(
-        self,
-        *,
-        event_id: str,
-        package_id: str,
-        source_run_id: str,
-        action: str,
-        actor: str,
-        label: str | None,
-        metadata: dict[str, Any] | None = None,
-        created_at: str = "",
-    ) -> None:
-        with self.connect() as conn:
-            conn.execute(
-                """
-                INSERT INTO hub_promotions(
-                    event_id, package_id, source_run_id, action, actor, label, metadata_json, created_at
-                )
-                VALUES (?, ?, ?, ?, ?, ?, ?, ?)
-                ON CONFLICT(event_id) DO UPDATE SET
-                    package_id = excluded.package_id,
-                    source_run_id = excluded.source_run_id,
-                    action = excluded.action,
-                    actor = excluded.actor,
-                    label = excluded.label,
-                    metadata_json = excluded.metadata_json
-                """,
-                (
-                    event_id,
-                    package_id,
-                    source_run_id,
-                    action,
-                    actor,
-                    label,
-                    json.dumps(metadata or {}),
-                    created_at,
-                ),
-            )
-
-    def get_hub_promotion_record(self, event_id: str) -> dict[str, Any] | None:
-        with self.connect() as conn:
-            row = conn.execute(
-                "SELECT * FROM hub_promotions WHERE event_id = ?",
-                (event_id,),
-            ).fetchone()
-            if row is None:
-                return None
-            return self._parse_hub_promotion_row(dict(row))
-
-    def list_hub_promotion_records(self) -> list[dict[str, Any]]:
-        with self.connect() as conn:
-            rows = conn.execute(
-                "SELECT * FROM hub_promotions ORDER BY created_at DESC"
-            ).fetchall()
-            return [self._parse_hub_promotion_row(dict(row)) for row in rows]
-
-    # ---- Monitor Conditions + Alerts (AC-209) ----
-
-    def insert_monitor_condition(self, condition: MonitorCondition) -> str:
-        """Persist a MonitorCondition. Returns the condition id."""
-        with self.connect() as conn:
-            conn.execute(
-                """
-                INSERT INTO monitor_conditions(id, name, condition_type, params_json, scope, active)
-                VALUES (?, ?, ?, ?, ?, ?)
-                """,
-                (
-                    condition.id,
-                    condition.name,
-                    str(condition.condition_type),
-                    json.dumps(condition.params),
-                    condition.scope,
-                    1 if condition.active else 0,
-                ),
-            )
-            return str(condition.id)
-
-    def list_monitor_conditions(
-        self,
-        *,
-        active_only: bool = True,
-        scope: str | None = None,
-    ) -> list[dict[str, Any]]:
-        """List monitor conditions with optional filters. Returns parsed params."""
-        query = "SELECT * FROM monitor_conditions WHERE 1=1"
-        params: list[Any] = []
-        if active_only:
-            query += " AND active = 1"
-        if scope is not None:
-            query += " AND scope = ?"
-            params.append(scope)
-        query += " ORDER BY created_at DESC"
-        with self.connect() as conn:
-            rows = conn.execute(query, params).fetchall()
-            results = []
-            for row in rows:
-                d = dict(row)
-                raw_params = d.pop("params_json", "{}")
-                d["params"] = json.loads(raw_params) if isinstance(raw_params, str) else {}
-                results.append(d)
-            return results
-
-    def count_monitor_conditions(
-        self,
-        *,
-        active_only: bool = True,
-        scope: str | None = None,
-    ) -> int:
-        """Count monitor conditions with optional filters."""
-        query = "SELECT COUNT(*) AS cnt FROM monitor_conditions WHERE 1=1"
-        params: list[Any] = []
-        if active_only:
-            query += " AND active = 1"
-        if scope is not None:
-            query += " AND scope = ?"
-            params.append(scope)
-        with self.connect() as conn:
-            row = conn.execute(query, params).fetchone()
-            return int(row["cnt"]) if row is not None else 0
-
-    def get_monitor_condition(self, condition_id: str) -> dict[str, Any] | None:
-        """Get a single monitor condition by id. Returns parsed params."""
-        with self.connect() as conn:
-            row = conn.execute(
-                "SELECT * FROM monitor_conditions WHERE id = ?",
-                (condition_id,),
-            ).fetchone()
-            if row is None:
-                return None
-            d = dict(row)
-            raw_params = d.pop("params_json", "{}")
-            d["params"] = json.loads(raw_params) if isinstance(raw_params, str) else {}
-            return d
-
-    def deactivate_monitor_condition(self, condition_id: str) -> bool:
-        """Deactivate a monitor condition. Returns True if found and updated."""
-        with self.connect() as conn:
-            conn.execute(
-                "UPDATE monitor_conditions SET active = 0 WHERE id = ?",
-                (condition_id,),
-            )
-            row = conn.execute("SELECT changes()").fetchone()
-            return bool(row[0] > 0) if row else False
-
-    def insert_monitor_alert(self, alert: MonitorAlert) -> str:
-        """Persist a MonitorAlert. Returns the alert id."""
-        with self.connect() as conn:
-            conn.execute(
-                """
-                INSERT INTO monitor_alerts(id, condition_id, condition_name, condition_type,
-                    scope, detail, payload_json, fired_at)
-                VALUES (?, ?, ?, ?, ?, ?, ?, ?)
-                """,
-                (
-                    alert.id,
-                    alert.condition_id,
-                    alert.condition_name,
-                    str(alert.condition_type),
-                    alert.scope,
-                    alert.detail,
-                    json.dumps(alert.payload),
-                    alert.fired_at,
-                ),
-            )
-            return str(alert.id)
-
-    def list_monitor_alerts(
-        self,
-        *,
-        condition_id: str | None = None,
-        scope: str | None = None,
-        limit: int = 100,
-        since: str | None = None,
-    ) -> list[dict[str, Any]]:
-        """List monitor alerts with optional filters. Returns parsed payload."""
-        query = "SELECT * FROM monitor_alerts WHERE 1=1"
-        params: list[Any] = []
-        if condition_id is not None:
-            query += " AND condition_id = ?"
-            params.append(condition_id)
-        if scope is not None:
-            query += " AND scope = ?"
-            params.append(scope)
-        if since is not None:
-            query += " AND fired_at >= ?"
-            params.append(since)
-        query += " ORDER BY fired_at DESC LIMIT ?"
-        params.append(limit)
-        with self.connect() as conn:
-            rows = conn.execute(query, params).fetchall()
-            results = []
-            for row in rows:
-                d = dict(row)
-                raw_payload = d.pop("payload_json", "{}")
-                d["payload"] = json.loads(raw_payload) if isinstance(raw_payload, str) else {}
-                results.append(d)
-            return results
-
-    def get_latest_monitor_alert(self, condition_id: str) -> dict[str, Any] | None:
-        """Return the newest alert for a condition, if one exists."""
-        alerts = self.list_monitor_alerts(condition_id=condition_id, limit=1)
-        return alerts[0] if alerts else None
