+++ autocontext/tests/test_module_size_limits.py
@@ -16,7 +16,7 @@ MAX_LINES = 800
 
 GRANDFATHERED: dict[str, int] = {
     # These are large but not yet split — tracked for future refactoring
-    "storage/sqlite_store.py": 1650,
+    "storage/sqlite_store.py": 1250,
     "storage/artifacts.py": 1300,
     "cli.py": 1300,
     "mcp/tools.py": 1500,
